Minimum to get GObject-introspected docs building
This adds a "GirScraper" scraper with its own minimal clean-html and
entries filters. All GObject-introspected documentation will be scraped
from HTML generated by g-ir-doc-tool's "devdocs" output format.

For now, clean-html does nothing, and entries adds a sidebar entry for
each node marked with an ".entry" CSS class. This entry appears in the
sidebar under the heading of the parent type.

This adds a `thor gir:generate` command which reads a GIR file and
generates a corresponding GirScraper-derived scraper in the directory
lib/docs/scrapers/gnome/generated/. Instead of relying on already-built
HTML documentation in a particular directory, as the rest of the scrapers
do, it builds DevDocs-specific HTML documentation with g-ir-doc-tool
during the initialization process of the scraper, in a temporary
directory.