Annotate deprecated entries in sidebar
We prepend deprecated entries in the sidebar with a warning sign
character. This makes them sort after everything else in the same class,
so it's more obvious that they're not the same as other class members; and
it also makes it easy to see at a glance that something is deprecated.