Generate versioned scraper names
It was originally a bad decision not to include the GIR's API version in
the scraper name; how else could you have two API versions of the same
documentation in parallel?

Now we append the API version (without dots) to the scraper name and
documentation slug. E.g., NameSpace-1.0 -> namespace10