Fix links to other GIRs
In g-ir-doc-tool we now mark links to other GIRs with a gir:// URI scheme
that DevDocs won't automatically process as an internal link. Therefore,
we have to add a fix_url option to resolve that link to another DevDocs
module.

We assume that all other referenced GIRs are present in DevDocs; if they
aren't, then you'll just get a not-found page when clicking on a dangling
link.