Change abbreviations
Most of the abbreviations are for docsets we're not using. Remove them
all except "js" and also add "gp" for GdkPixbuf, which is annoying to
type.

We create a new ALIASES variable in the GNOME overrides instead of
editing the first definition, so that we can avoid merge conflicts with
upstream DevDocs.

(Ported to GNOME overrides, and from CoffeeScript to JS, by Evan Welsh.)