Branding and help refer to GNOME
Now that this is online and people are starting to use it, add some GNOME
branding. Previously the DevDocs "thanks for downloading" developer
message was shown on startup; now we let people know that this is GNOME
JavaScript API documentation and that it's a modified instance of DevDocs.
Slight wording changes throughout.

We temporarily get rid of the News page, since most of the news is about
irrelevant docsets. Once we have our own news, we'll bring it back.