Adapt Docker image to Fedora 29 and rbenv
NetworkManager-glib-devel, clutter-gst, clutter-gst2, vte3-devel, and
the old WebKit packages are no longer in Fedora 29. This means that we
lose the cluttergst10, cluttergst20, vte290, webkit10, and webkit30
docsets. The gstfft10 docset apparently no longer exists;
networkmanager10 and nmclient10 are replaced by nm10; and tracker10,
trackercontrol10, and trackerminer10 are replaced by tracker20,
trackercontrol20, and trackerminer20 respectively.

Additionally, we can no longer use rvm to build the needed version of
Ruby, as it requires adding the user to the rvm group, which would
require logging out and logging in again in the Dockerfile. Instead we
use rbenv which uses a different mechanism.